Added doc strings
