+++ RoeiJeRot.View.Wpf/Views/Windows/MainWindow.xaml.cs
@@ -3,6 +3,7 @@ using System.ComponentModel;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Input;
+using RoeiJeRot.View.Wpf.Logic;
 using RoeiJeRot.View.Wpf.ViewModels;
 
 namespace RoeiJeRot.View.Wpf.Views.Windows
