+++ RoeiJeRot.View.Wpf/Views/Windows/ReservationOverviewWindow.xaml.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Windows;
 using System.Windows.Controls;
 using RoeiJeRot.Logic.Services;
+using RoeiJeRot.View.Wpf.Logic;
 using RoeiJeRot.View.Wpf.ViewModels;
 
 namespace RoeiJeRot.View.Wpf.Views.Windows
